CryptoConnect/backend/requirements.txt
fastapi==0.115.2
starlette==0.40.0
uvicorn[standard]==0.30.3
motor==3.7.1
beanie==2.0.0
pydantic==2.11.7
pydantic-settings==2.10.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
xrpl-py==4.3.0
python-dotenv==1.1.1
bcrypt==4.0.1
pymongo==4.13.2